+++ CHANGELOG.md
@@ -5,10 +5,14 @@ All notable changes to **open-quill** are documented here.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2026.2.2] — 2026-06-11
+## [2026.2.2] — 2026-06-12
+
+### Added
+- Chats Tab: allowing you to view all your previous chats in one spot!
 
 ### Changed
 - Updated **ALL** dependencies to the **Latest** version.
+- Refactored `Models` menu in Admin Panel.
 
 ### Fixed
 - Fix `.reasoning-body` margin.